Add ROM::Model::Form that acts kinda like an AR object
I've gone through pain and suffering while implementing this but here it
goes. This encapsulates defining params and validations within a single
class that can be used in a similar way as an AR object with form_for.

It allows you to specify which ROM commands a given form requires and
expects you to implement `commit` method where you do your stuff with
commands. It will provide param values and potential errors too which is
handy in the view layer.

This is a start, not sure where this is going but I will try to use it
and see if it makes sense in the long run.