form.errors may return a blank ActiveModel::Errors object
A number of form-building libraries, such as simple_form and formtasic,
depend on the active-model errors interface.  They expect Errors to
always be present, never nil,  and to return an empty array for specific
attribute errors.

This bounces forward an empty errors object, when no errors are found in
the result.

Note:  I've only added unit tests for the base, pre-save case.  the
integration specs seem to have driven the change to the success? method
without needing any additional specs