Use HashWithIndifferentAccess for params
Oh boy, this is an indication that form for a new object should have
different behavior than a form for existing object. Why? Because a new
object form receives attributes from rack env (stringified hash) and a
form for an existing object has symbolized hash params.

I'm gonna leave it for now like this, it kinda works, doesn't make me
happy but oh well, it's Rails integration so it's not gonna be pretty.