allow the reader builder to skip already defined methods
This is going needed because we'll be running through the input method a
couple of times now, and some methods will have already been assigned by
a parent class.

This is the non-destructive version of a reassignment, and may be less
surprising.  The current implementation may raise an error depending on
_where_ in the file the input method is defined, relative to a duplicate
reader method, so not raising an error may be more consistent.