Convert provided parameters to an unsafe hash.
Rails5 has changed the hierarchy of the `Parameters` class, presumably
to better enforce safe_parameter filtering.  It now throws a warning
when `to_hash` is called, preferring instead that `#to_h` or
`#to_unsafe_h` is used.

In our case, we actually _want_ `#to_unsafe_h`, because our form objects
take the place of the `strong_parameter` filtering that is applied in
the `to_h` method.