first cut at a github build workflow to replace travis
based in some part on the ci for rom core.

This matrixes over all of the exepected Rails versions and ruby
versions, same as Travis does.  I'll want to look at replicating the
git@master inclusion for ROM and ROM-SQL as well, to attempt a guard
against sudden breakage, but will handle that separately.